Changed the types

+++ app/src/main/cpp/Actor.cpp
@@ -1,7 +1,7 @@
 #include "Actor.h"
 #include "GameObjectManager.h"
 
-Actor::Actor(unsigned int id, size_t componentsSize)
+Actor::Actor(uint32_t id, uint32_t componentsSize)
 	:	id(id), components(componentsSize)
 {
 }
@@ -10,8 +10,8 @@ void Actor::update(float dt)
 {
 	for (auto it = components.begin(); it != components.end();)
 	{
-		uint compSize = *((uint*) it);
-		it += sizeof(uint);
+		uint32_t compSize = *((uint32_t*) it);
+		it += sizeof(uint32_t);
 
 		((Component*)(it))->update(dt, this);
 
@@ -19,12 +19,12 @@ void Actor::update(float dt)
 	}
 }
 
-int Actor::getComponentIndex(uint componentId) const
+int32_t Actor::getComponentIndex(uint32_t componentId) const
 {
 	for(auto it = components.begin(); it != components.end();)
 	{
-		uint compSize = *((uint*) it);
-		it += sizeof(uint);
+		uint32_t compSize = *((uint32_t*) it);
+		it += sizeof(uint32_t);
 
 		if(((Component*)(it))->getId() == componentId)
 			return (int)(it - components.begin());
