+++ app/src/main/cpp/Texture.cpp
@@ -8,13 +8,13 @@
 
 static void callbackReadPng(png_structp pngStruct, png_bytep data, png_size_t size) {
 	Ifstream* asset = ((Ifstream*)png_get_io_ptr(pngStruct));
-	asset->read(data, size);
+	asset->read(data, (uint32_t)size);
 }
 
 static void callbackReadTransform(png_structp ptr, png_row_infop rowInfo, png_bytep data)
 {
     assert(rowInfo->bit_depth == 8); // 1 byte, uchar
-    assert(rowInfo->rowbytes == (rowInfo->width * sizeof(uint)));
+    assert(rowInfo->rowbytes == (rowInfo->width * sizeof(uint32_t)));
     assert(rowInfo->color_type == PNG_COLOR_TYPE_RGB || rowInfo->color_type == PNG_COLOR_TYPE_RGBA);
 
     if(rowInfo->color_type == PNG_COLOR_TYPE_RGBA)
@@ -22,13 +22,13 @@ static void callbackReadTransform(png_structp ptr, png_row_infop rowInfo, png_by
         assert(rowInfo->channels == 4); //RGBA
         assert(rowInfo->pixel_depth == (4 * 8));
 
-        for(int i = 0; i < rowInfo->rowbytes; i += 4)
+        for(int32_t i = 0; i < rowInfo->rowbytes; i += 4)
         {
             float alpha = data[i + 3] / 255.0f;
 
-            data[i + 0] = (uchar) (data[i + 0] / 255.0f * alpha * 255.0f);
-            data[i + 1] = (uchar) (data[i + 1] / 255.0f * alpha * 255.0f);
-            data[i + 2] = (uchar) (data[i + 2] / 255.0f * alpha * 255.0f);
+            data[i + 0] = (uint8_t) (data[i + 0] / 255.0f * alpha * 255.0f);
+            data[i + 1] = (uint8_t) (data[i + 1] / 255.0f * alpha * 255.0f);
+            data[i + 2] = (uint8_t) (data[i + 2] / 255.0f * alpha * 255.0f);
         }
     }
 }
@@ -135,7 +135,7 @@ bool Texture::loadFromFile(const String & filename)
 	png_byte* image = new png_byte[rowSize * height];
 	png_bytep* rowPtrs = new png_bytep[height];
 
-	for (uint i = 0, otherI = height - 1; i < height; ++i, --otherI)
+	for (uint32_t i = 0, otherI = height - 1; i < height; ++i, --otherI)
 	{
 		rowPtrs[otherI] = &image[i * rowSize];
 	}
@@ -201,13 +201,13 @@ Texture::operator bool() const
 	return (width != 0);
 }
 
-void Texture::bind(int slot) const
+void Texture::bind(int32_t slot) const
 {
 	CallGL(glActiveTexture(GL_TEXTURE0 + slot));
 	CallGL(glBindTexture(GL_TEXTURE_2D, texture));
 }
 
-Texture::Texture(const void* buffer, int width, int height, GLint internalFormat)
+Texture::Texture(const void* buffer, int32_t width, int32_t height, GLint internalFormat)
 {
 	CallGL(glGenTextures(1, &texture));
 	CallGL(glActiveTexture(GL_TEXTURE0));
